Break circular dep on XliffInterface
